+++ libraries/TCA9555/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.4] - 2022-12-29
+- Fix #7 missing public:
+- add unit test for TCA9535 constructor
+- add examples for TCA9535
+- update examples
+- add **getAddress()** (debugging)
+- update readme.md
+- update GitHub actions -> v3
+
+
 ## [0.1.3] - 2022-11-25
 - Add RP2040 support to build-CI.
 - Add CHANGELOG.md